poo, global test_model, global_prediction, comments on package
